fix: patch behavior around schema upgrades (#66)
This is all due to the can of worms that was opened up with #46, revealing a fundamental flaw in how schema upgrades were (not) handled. Since all pointers in this version of the library are equivalent to "builders" in the reference implementation, now all attempts to get a list or struct that is too small for its content will shallow-copy and resize them appropriately before returning the pointer. Downgrades are still handled the right way by not touching the underlying data at all. (It's already large enough to hold everything, and all the existing offsets must still be valid.)

Closes #46.