pack messages segment by segment (#70)
This resolves a crucial flaw in the way messages were packed; now each segment is packed separately with no need for unpacking to maintain state across each segment. The packing was "valid" (in a sense) before, but it didn't work with the C++ implementation which is big time no bueno.