fix: upgrade outdated packages, revert to makefile (#159)
Fixes some long-standing issues with the build related to outdated dependencies, and [a return to using a Makefile](https://www.olioapps.com/blog/the-lost-art-of-the-makefile/). Given the number of breaking changes in dependencies it should be assumed this may introduce bugs, despite a healthy test suite (now running on GitHub Actions, :wave: Travis CI).